[feat] Change routing to /cluster/NAME/...
Previously, the cluster name was hard-coded. This commit moves all of the routes around
(so we don't have to change resolver) and allows querying differnt clusters. At any time,
we can have only one active cluster. The dashboard will not use /cluster/... subtree, so
support for multiple clusters does not have to be included.